feat: non-existent book delete treatment

+++ src/domain/use-cases/delete-book.ts
@@ -1,10 +1,13 @@
-import { DeleteBook as DeleteBookInterface } from '@/domain/contracts/repo'
+import { BookNotExist } from '@/domain/errors'
+import { DeleteBook as DeleteBookInterface, LoadBookBySbn } from '@/domain/contracts/repo'
 
-type Setup = (bookRepo: DeleteBookInterface) => DeleteBook
+type Setup = (bookRepo: DeleteBookInterface & LoadBookBySbn) => DeleteBook
 type Input = { sbn: string }
 
 export type DeleteBook = (input: Input) => Promise<void>
 
 export const setupDeleteBook: Setup = (bookRepo) => async ({ sbn }) => {
+  const book = await bookRepo.loadBySbn({ sbn })
+  if (book === undefined) throw new BookNotExist()
   await bookRepo.delete({ sbn })
 }
